Add GeomPicker::GetSetValidGeom()
This validates the stored m_GeomIDChoice and returns it.

If the stored m_GeomIDChoice is not valid, it either chooses the first
from the list of valid choices -- or (if there are none) it sets the
empty string.

After this call, m_GeomIDChoice will either be valid or empty -- and
that value will be returned.