Improve TBndBox::Intersect() to allow self-intersection
Mesh self-intersection was not previously supported.  These changes
address problems that arise when performing self intersection.

The first check skips simple self-intersection of a triangle.

The second check skips tris that share an edge.  These could intersect,
but it would always be a pathological coplanar case that should not
occur.

At this point, TMesh triangles are made of overlapping duplicated nodes,
so checking if the tris share an edge must be done with a distance
check (in ttri::ShareEdge()).

Neither of these tests should trigger in the prior CompGeom workflow.

The final test rejects intersection segments that fall outside of either
triangle.  This should never happen, but does when two tris share a node
and have a colinear edge.

This test should also be safe in all cases.

An alternate fix would be to use a different tri-tri intersection routine
that does not give the errant results in this case.