Add GetUSkip() method
GetUSkip (and GetRootCluster / GetTipCluster) need to be indexed
according to the original number of surface segments -- i.e. before
any changes incurred by GeomEngine stuff.

USkip is only used to skip the U caps when generating DegenGeom.

So, after this change, caps may still not be properly skipped for Engine
DegenGeoms, but the common case of regular Engine geoms should be fixed.