Simplify Geom::ApplySymm() to use information stored in UpdateSymmAttach()
Copying, flipping normals, and transforming surfaces used to happen in
UpdateSymmAttach().  When it was separated out to ApplySymm()
(before it was made a templated method), most of the logic was
kept in place and then stripped down.

However, all the information required to perform the same function
was stored in m_FlipNormalVec, m_TransMatVec, and m_MainSurfIndxVec.

This change uses that stored information to replace the messy logic.