OpenVSP 3.44.3
This is a bugfix release made by back porting applicable fixes from
the 3.45.X branch.  Update to this version if you can't update to
3.45.X for some reason.  That said, 3.45.X has been going very well
and I do not anticipate keeping this 3.44.X branch active for long.

Although a bugfix release, a couple small features squeezed their way
in to help smooth some of the rough edges in the big new features
introduced in 3.44.0.

The new Geometry Analysis feature gained an online help page to help
explain the input requirements for each analysis.  The analyses also
received some error checking to notify the user when those requirements
are not met.

The most visible change in this version will be the layout of the
main OpenVSP screen.  Over the years, screens have gotten bigger and
wider -- and while most have landscape orientation, some are portrait.
The default size of the main screen has been increased to better take
advantage of the screen.

Features:
 - Geometry Analysis Online Help
 - Error handling for Geometry Analysis
 - Adjust Main Screen sizing

Bug Fixes:
 - Expose AuxGeom XSec Attributes
 - Fix bug with tessellation of EngineGeom
 - Fix CFDMesh crash tied to model scale.
 - Fix reading STL files on Windows with non-native line endings.
 - Fix Bogie containers not appearing in search
 - Fix CG envelope calculations & visualization
 - Clean up GearScreen
 - Improve matrix attribute format string