Calculate FEA Beam cross section property values.
The cross sectional area was not calculated, which led to OpenVSP's
mass calculation being wrong.

The other properties are not strictly needed, but are nice to have if
one wants to initialize section properties with a defined section but
then switch to a GENERAL section and have all the values already set.