Use VspSlider instead of Fl_Slider to fix deactivated behavior
A VspSlider was created to replace the supplied Fl_Slider.  Although it
was AddWidget()'d in Slider::Init(), all the more complex Init() methods
would ClearAllWidgets() and then AddWidget() to the widgets known at thier
level.

This resulted in the underlying VspSlider not being in the m_WidgetVec, which
was used for Activate() and Deactivate().

Consequently, sliders were immune to deactivation.