Invert Deactivate() / Activate() logic in HingeScreen::Update()
This fixes a problem introduced by d3adcd370cd.  Previously, Deactivate()
never worked on Sliders.  That commit fixed this.

However, when a Slider is Deactivate()'ed in the middle of a slide, the slide
stops working.

There is a common pattern in the GUI where GuiDevices are
Activate()/Deactivate()'ed en mass, then selectively Deactivate()/Activate()'ed
to get the desired setting.

If Sliders are involved, we should always Activate() first, then Deactivate()
as needed.

This change fixes this behavior for Hinge Sliders.  More similar changes
are likely needed.