Add m_SkipCurveMods flag to XSecCurve and use to skip curve modifications
The curve modifications would always result in a closed curve.  When using
an XSecCurve based SubSurface, this is not desired.  Further, the Parms needed
to set the curve modifications are not exposed.

Rather than set up an elaborate open curve capability, this flag is a bit of a
hack to just skip all the modifications entirely.